+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/style/standard/MapboxStandardStyle.kt
@@ -10,7 +10,7 @@ import com.mapbox.maps.extension.compose.MapboxMapComposable
 import com.mapbox.maps.extension.compose.style.GenericStyle
 import com.mapbox.maps.extension.compose.style.MapboxStyleComposable
 import com.mapbox.maps.extension.compose.style.atmosphere.generated.AtmosphereState
-import com.mapbox.maps.extension.compose.style.projection.Projection
+import com.mapbox.maps.extension.compose.style.projection.generated.Projection
 import com.mapbox.maps.extension.compose.style.slotsContent
 import com.mapbox.maps.extension.compose.style.styleImportsConfig
 import com.mapbox.maps.extension.compose.style.terrain.generated.TerrainState
@@ -23,16 +23,22 @@ import com.mapbox.maps.extension.style.expressions.generated.Expression
  */
 @Immutable
 @MapboxExperimental
-public data class LightPreset(public val value: Value) {
+public data class LightPresetValue(public val value: Value) {
   /**
-   * Construct the LightPreset with [String].
+   * Construct the [LightPresetValue] with [String].
    */
   public constructor(value: String) : this(Value(value))
 
   /**
-   * Construct the LightPreset with [Mapbox Expression](https://docs.mapbox.com/style-spec/reference/expressions/).
+   * Construct the [LightPresetValue] with [Mapbox Expression](https://docs.mapbox.com/style-spec/reference/expressions/).
    */
   public constructor(expression: Expression) : this(expression as Value)
+
+  /**
+   * True if the this value is not [INITIAL]
+   */
+  internal val notInitial: Boolean
+    get() = this !== INITIAL
 
   /**
    * Public companion object.
@@ -41,33 +47,45 @@ public data class LightPreset(public val value: Value) {
     internal const val NAME: String = "lightPreset"
 
     /**
-     * Default value for [LightPreset], setting default will result in restoring the property value defined in the style.
+     * Use this constant to signal that no property should be set to the Maps engine.
+     * This is needed because sending nullValue resets the value of the property to the default one
+     * defined by the Maps engine, which results in overriding the value from the loaded style.
+     * Moreover, we set a custom String to differentiate it from [DEFAULT], otherwise things
+     * like [kotlinx.coroutines.flow.Flow] or [androidx.compose.runtime.MutableState] won't be able
+     * to differentiate them because they use [equals].
      */
-    public val default: LightPreset = LightPreset(Value.nullValue())
+    @JvmField
+    internal val INITIAL: LightPresetValue = LightPresetValue(Value.valueOf("LightPresetValue.INITIAL"))
+
+    /**
+     * Default value for [LightPresetValue], setting default will result in restoring the property value defined in the style.
+     */
+    @JvmField
+    public val DEFAULT: LightPresetValue = LightPresetValue(Value.nullValue())
 
     /**
      * The light preset for "day".
      */
     @JvmField
-    public val DAY: LightPreset = LightPreset("day")
+    public val DAY: LightPresetValue = LightPresetValue("day")
 
     /**
      * The light preset for "night".
      */
     @JvmField
-    public val NIGHT: LightPreset = LightPreset("night")
+    public val NIGHT: LightPresetValue = LightPresetValue("night")
 
     /**
      * The light preset for "dusk".
      */
     @JvmField
-    public val DUSK: LightPreset = LightPreset("dusk")
+    public val DUSK: LightPresetValue = LightPresetValue("dusk")
 
     /**
      * The light preset for "dawn".
      */
     @JvmField
-    public val DAWN: LightPreset = LightPreset("dawn")
+    public val DAWN: LightPresetValue = LightPresetValue("dawn")
   }
 }
 
@@ -78,7 +96,7 @@ public data class LightPreset(public val value: Value) {
  * @param topSlot The content to be set to the top slot of the Mapbox Standard style.
  * @param middleSlot The content to be set to the middle slot of the Mapbox Standard style.
  * @param bottomSlot The content to be set to the bottom slot of the Mapbox Standard style.
- * @param lightPreset The [LightPreset] settings of the Mapbox Standard Style, available lightPresets including "day", "night", "dawn", "dusk".
+ * @param lightPreset The [LightPresetValue] settings of the Mapbox Standard Style, available lightPresets including "day", "night", "dawn", "dusk".
  * @param projection The projection to be set to the map. Defaults to [Projection.default] meaning that projection value is taken from the Standard style definition.
  * @param atmosphereState The atmosphere to be set to the map By default, no changes to the current atmosphere.
  * @param terrainState The terrain to be set to the map. Defaults to initial state meaning no custom terrain is added, default value is taken from Standard style definition.
@@ -90,10 +108,10 @@ public fun MapboxStandardStyle(
   topSlot: (@Composable @MapboxMapComposable () -> Unit)? = null,
   middleSlot: (@Composable @MapboxMapComposable () -> Unit)? = null,
   bottomSlot: (@Composable @MapboxMapComposable () -> Unit)? = null,
-  lightPreset: LightPreset = LightPreset.default,
-  projection: Projection = Projection.default,
+  lightPreset: LightPresetValue = LightPresetValue.INITIAL,
+  projection: Projection = Projection.INITIAL,
   atmosphereState: AtmosphereState = remember { AtmosphereState() },
-  terrainState: TerrainState = TerrainState.initial,
+  terrainState: TerrainState = TerrainState.INITIAL,
 ) {
   GenericStyle(
     style = Style.STANDARD,
@@ -103,8 +121,10 @@ public fun MapboxStandardStyle(
       bottomSlot?.let { slot("bottom", it) }
     },
     styleImportsConfig = styleImportsConfig {
-      importConfig(importId = "basemap") {
-        config(LightPreset.NAME, lightPreset.value)
+      if (lightPreset.notInitial) {
+        importConfig(importId = "basemap") {
+          config(LightPresetValue.NAME, lightPreset.value)
+        }
       }
     },
     projection = projection,